refactor: move funding source linking to GroupsMixin
update tests as well
