+++ littlepay/api/groups.py
@@ -2,6 +2,7 @@ from dataclasses import dataclass
 from typing import Generator
 
 from littlepay.api import ClientProtocol
+from littlepay.api.funding_sources import FundingSourcesMixin
 
 
 @dataclass
@@ -20,6 +21,10 @@ class GroupsMixin(ClientProtocol):
         """Endpoint for concession groups. Optionally provide a group_id for a group-specific endpoint."""
         return self._make_endpoint(self.CONCESSION_GROUPS, group_id, *parts)
 
+    def concession_group_funding_source_endpoint(self, group_id: str) -> str:
+        """Endpoint for a concession group's funding sources."""
+        return self.concession_groups_endpoint(group_id, FundingSourcesMixin.FUNDING_SOURCES)
+
     def create_concession_group(self, group_label: str) -> dict:
         """Create a new concession group."""
         endpoint = self.concession_groups_endpoint()
@@ -36,3 +41,9 @@ class GroupsMixin(ClientProtocol):
         """Remove an existing concession group."""
         endpoint = self.concession_groups_endpoint(group_id)
         return self._delete(endpoint)
+
+    def link_concession_group_funding_source(self, group_id: str, funding_source_id: str) -> dict:
+        """Link a funding source to a concession group."""
+        endpoint = self.concession_group_funding_source_endpoint(group_id)
+        data = {"id": funding_source_id}
+        return self._post(endpoint, data, dict)
